diary ui done, preparation for add bloc state managment table calendar

+++ lib/features/diary/domain/entities/ingredient_entity.dart
@@ -6,9 +6,9 @@ class IngredientEntity extends Equatable{
   final int calories;
   final String name;
   final int weight;
-  final double proteins;
-  final double carbohydrates;
-  final double fats;
+  final int proteins;
+  final int carbohydrates;
+  final int fats;
   final String urlPicture;
   IngredientEntity({required this.id, required this.calories, required this.name, required this.weight, required this.proteins, required this.carbohydrates, required this.fats, required this.urlPicture});
